wifi: Do not set any packet filtering
It seems that notifications are not getting through over WiFi when wlan
is suspended because of a mismatch in CAF's default packet filtering
configuration compared to OOS's (see wlan_hdd_packet_filter_rules.h).
Since CAF does not enable this by default in Cobalt, we shall disable
it as well.

Change-Id: I27b547569953553ea4ee7c04d7b0716b9b1de300
Signed-off-by: joshuous <joshuous@gmail.com>
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>